+++ src/lib.rs
@@ -10,7 +10,10 @@ pub use self::{
     row::{Row, RowOwned, RowRead, RowWrite},
 };
 use self::{error::Result, http_client::HttpClient};
+
+#[doc = include_str!("row_derive.md")]
 pub use clickhouse_derive::Row;
+
 use std::{collections::HashMap, fmt::Display, sync::Arc};
 
 pub mod error;
@@ -37,6 +40,12 @@ mod rowbinary;
 mod ticks;
 
 /// A client containing HTTP pool.
+///
+/// ### Cloning behavior
+/// Clones share the same HTTP transport but store their own configurations.
+/// Any `with_*` configuration method (e.g., [`Client::with_option`]) applies
+/// only to future clones, because [`Client::clone`] creates a deep copy
+/// of the [`Client`] configuration, except the transport.
 #[derive(Clone)]
 pub struct Client {
     http: Arc<dyn HttpClient>,
@@ -517,5 +526,19 @@ mod client_tests {
         assert!(!client.validation);
         let client = client.with_validation(true);
         assert!(client.validation);
+    }
+
+    #[test]
+    fn it_does_follow_previous_configuration() {
+        let client = Client::default().with_option("async_insert", "1");
+        assert_eq!(client.options, client.clone().options,);
+    }
+
+    #[test]
+    fn it_does_not_follow_future_configuration() {
+        let client = Client::default();
+        let client_clone = client.clone();
+        let client = client.with_option("async_insert", "1");
+        assert_ne!(client.options, client_clone.options,);
     }
 }