+++ tests/it/main.rs
@@ -31,7 +31,7 @@ use std::sync::LazyLock;
 
 macro_rules! assert_panic_on_fetch_with_client {
     ($client:ident, $msg_parts:expr, $query:expr) => {
-        use futures::FutureExt;
+        use futures_util::FutureExt;
         let async_panic =
             std::panic::AssertUnwindSafe(async { $client.query($query).fetch_all::<Data>().await });
         let result = async_panic.catch_unwind().await;
@@ -48,7 +48,7 @@ macro_rules! assert_panic_on_fetch_with_client {
 
 macro_rules! assert_panic_on_fetch {
     ($msg_parts:expr, $query:expr) => {
-        use futures::FutureExt;
+        use futures_util::FutureExt;
         let client = get_client();
         let async_panic =
             std::panic::AssertUnwindSafe(async { client.query($query).fetch_all::<Data>().await });